Finish implementing logic for author submission approval/rejection

+++ backend/src/main/java/com/github/sveldevorls/readtogether/submission/service/AuthorSubmissionService.java
@@ -1,11 +1,13 @@
 package com.github.sveldevorls.readtogether.submission.service;
 
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.github.sveldevorls.readtogether.submission.entity.AuthorSubmission;
 import com.github.sveldevorls.readtogether.submission.entity.AuthorSubmissionMapper;
 import com.github.sveldevorls.readtogether.author.service.AuthorService;
+import com.github.sveldevorls.readtogether.common.entity.ReviewStatus;
 import com.github.sveldevorls.readtogether.common.exception.ResourceNotFoundException;
 import com.github.sveldevorls.readtogether.submission.dao.AuthorSubmissionDao;
 import com.github.sveldevorls.readtogether.submission.dto.AuthorSubmissionResponse;
@@ -49,46 +51,68 @@ public class AuthorSubmissionService {
     }
 
     @Transactional
-    public AuthorSubmissionResponse approveSubmission(int id, String reviewerComment) {
+    public AuthorSubmissionResponse approveSubmission(
+            int submissionId,
+            int reviewerId,
+            String reviewerComment) {
 
         // Fetch original author id for the submission
         int authorId = authorSubmissionDao
-                .getAuthorIdById(id)
+                .getAuthorIdById(submissionId)
                 .orElseThrow(() -> new ResourceNotFoundException());
 
         // Update submission review status
-        authorSubmissionDao.updateReviewStatusById(id, "approved");
-        authorSubmissionDao.updateReviewerCommentById(id, reviewerComment);
+        int submissionRows = authorSubmissionDao.updateReviewById(
+                submissionId,
+                ReviewStatus.approved,
+                reviewerId,
+                reviewerComment);
+
+        if (submissionRows == 0)
+            throw new ResourceNotFoundException("Submission was not found, try refreshing the page");
+        if (submissionRows > 1)
+            throw new DataIntegrityViolationException("Multiple submissions were found when only one was expected");
 
         // Mark the corresponding author as approved
         authorService.approveAuthor(authorId);
 
         // Fetch the newly updated submission response
         AuthorSubmissionResponse response = authorSubmissionDao
-                .getSubmissionResponseById(id)
+                .getSubmissionResponseById(submissionId)
                 .orElseThrow(() -> new ResourceNotFoundException());
 
         return response;
     }
 
     @Transactional
-    public AuthorSubmissionResponse rejectSubmission(int id, String reviewerComment) {
+    public AuthorSubmissionResponse rejectSubmission(
+            int submissionId,
+            int reviewerId,
+            String reviewerComment) {
 
         // Fetch original author id for the submission
         int authorId = authorSubmissionDao
-                .getAuthorIdById(id)
+                .getAuthorIdById(submissionId)
                 .orElseThrow(() -> new ResourceNotFoundException());
 
         // Update submission review status
-        authorSubmissionDao.updateReviewStatusById(id, "approved");
-        authorSubmissionDao.updateReviewerCommentById(id, reviewerComment);
+        int rows = authorSubmissionDao.updateReviewById(
+                submissionId,
+                ReviewStatus.rejected,
+                reviewerId,
+                reviewerComment);
+
+        if (rows == 0)
+            throw new ResourceNotFoundException("Submission was not found, try refreshing the page");
+        if (rows > 1)
+            throw new DataIntegrityViolationException("Multiple submissions were found when only one was expected");
 
         // Remove the rejected author
         authorService.deleteAuthor(authorId);
 
         // Fetch the newly updated submission response
         AuthorSubmissionResponse response = authorSubmissionDao
-                .getSubmissionResponseById(id)
+                .getSubmissionResponseById(submissionId)
                 .orElseThrow(() -> new ResourceNotFoundException());
 
         return response;
